python/requirements.txt
Pillow==9.3.0
matplotlib==3.7.1
